requirements.txt
pandas==2.2.3
numpy==2.0.2
faiss-cpu==1.10.0
sentence-transformers==4.1.0
llama-cpp-python==0.3.8
python-dotenv==1.1.0
torch==2.6.0
Flask==3.1.0
flask_cors==5.0.1
gunicorn==23.0.0
redis==6.0.0
Flask-Limiter==3.11.0